Backslash escape tilde characters in README.md
Github was rendering a large portion of the Quick Start section with
strikethroughs. I wasn't reading carefully, so I skimmed it and thought it said
to install in ~/.vim/plugin or ~/.config/nvim/plugin, which is actually the
opposite of what the paragraph says to do.

Hopefully, this should help prevent anyone else from misreading the README.md on
Github.